requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
langdetect>=1.0.9
google-generativeai>=0.3.0
python-dotenv>=1.0.0
notebook>=6.5.2
